# newservice/api-spec.yaml
openapi: 3.0.0
info:
  title: E-Commerce Order Management API
  description: API для управления заказами интернет-магазина с микросервисной архитектурой
  version: 1.0.0
  contact:
    name: Дипломный проект
    email: student@university.edu
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad request"
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin, manager]
          example: "user"
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        isActive:
          type: boolean
          example: true
...
    OrderItem:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          example: 101
        productName:
          type: string
          example: "Ноутбук ASUS"
        quantity:
          type: integer
          minimum: 1
          example: 2
        price:
          type: number
          example: 75000.50
        total:
          type: number
          example: 150001.00
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1001
        orderNumber:
          type: string
          example: "ORD-2024-0001"
        userId:
          type: integer
          example: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          example: 150001.00
        status:
          type: string
          enum: [pending, processing, paid, shipped, delivered, cancelled]
          example: "pending"
        paymentMethod:
          type: string
          enum: [card, cash, balance]
          example: "card"
        paymentId:
          type: string
          example: "pay_123456"
        shippingAddress:
          type: string
          example: "ул. Ленина, д. 1, кв. 10"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deliveryDate:
          type: string
          format: date-time
          nullable: true
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          example: "securePassword123"
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          format: password
          minLength: 6
          example: "securePassword123"
    CreateOrderRequest:
      type: object
      required:
        - items
        - shippingAddress
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: integer
                example: 101
              quantity:
                type: integer
                minimum: 1
                example: 2
        shippingAddress:
          type: string
          example: "ул. Ленина, д. 1, кв. 10"
        paymentMethod:
          type: string
          enum: [card, cash, balance]
          default: "card"
        promoCode:
          type: string
          example: "SUMMER2024"
          nullable: true
...
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        user:
          $ref: '#/components/schemas/User'
        tokens:
          type: object
          properties:
            accessToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIs..."
            refreshToken:
              type: string
              example: "eyJhbGciOiJIUzI1NiIs..."
            expiresIn:
              type: integer
              example: 3600
    OrdersListResponse:
      type: object
      properties:
        total:
          type: integer
          example: 42
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
  parameters:
    orderIdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: ID заказа
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: Количество записей на странице
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Смещение от начала

# Аутентификация и пользователи
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      description: Создает новую учетную запись пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      description: Аутентификация пользователя и получение токенов
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Аккаунт заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /orders:
    post:
      tags:
        - Orders
      summary: Создание нового заказа
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  order:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Ошибка в данных
        '401':
          description: Не авторизован
        '503':
          description: Сервис временно недоступен

    get:
      tags:
        - Orders
      summary: Получение списка заказов
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, paid, shipped, delivered, cancelled]
          description: Фильтр по статусу
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersListResponse'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Получение заказа по ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/orderIdParam'
      responses:
        '200':
          description: Детали заказа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '403':
          description: Нет доступа к заказу
        '404':
          description: Заказ не найден
...
  /internal/verify/{token}:
    get:
      tags:
        - Internal
      summary: Проверка токена (внутренний)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'